# openapi/api.yaml
openapi: 3.0.0
info:
  title: Events API
  version: 1.0.0
  description: API for retrieving and managing events.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /auth/refresh:
    post:
      summary: Refresh an authentication token
      operationId: refreshToken
      responses:
        "200":
          description: Token refreshed successfully, returns a new token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        "400":
          description: Invalid refresh token
...